Fix `test_SIGHUP_tty` under pytest
Since pytest patches stdin unless `-s` is used, it caused
`os.isatty(sys.stdin.fileno())` to report that the process is
daemonized. Pytest does not have a mechanism to temporarily unpatch
stdin so this change explicitly creates a Pseudo-TTY (PTY) and binds
it to subprocesses created by the test helper called `Process`.

Refs:
* https://stackoverflow.com/a/43012138/595220
* https://stackoverflow.com/q/41274719/595220