Replace occurances for www.cherrypy.org to www.cherrypy.dev
Post the domain migration mentioned
 https://github.com/python3statement/python3statement.github.io/pull/294#issue-2382318887
to the www.cherrypy.dev it was required to change the occurrences
accross the codebase as well.

Along with that noticed the license was still redirecting
to the domain cherrypy.org changed it to cherrypy.dev.

Also one thing came to the notice while replacing that there
are tickets
- http://www.cherrypy.org/ticket/581
- http://www.cherrypy.org/ticket/751

which are no longer at the domain mentioned above upon discussion
change those links to the github.com/issues